requirements.txt
# --- Core Web Framework ---
fastapi==0.115.2
uvicorn==0.30.3
python-dotenv==1.0.1
requests>=2.31.0

# --- Voice, Audio & Speech ---
speechrecognition==3.10.4
sounddevice>=0.4.6
simpleaudio>=1.0.4
wavio>=0.0.7
pyaudio>=0.2.14
gtts>=2.5.3

# --- Twilio / Exotel Integration ---
twilio==8.10.0
websockets>=11.0.3

# --- Google Sheets Integration ---
gspread>=6.2.1
google-auth>=2.29.0

# --- AI / LLM Framework ---
langchain>=0.3.0
langgraph>=0.5.0
openai>=1.12.0
# SarvamAI (for Text-to-Speech)
sarvamai~=0.1.15

# --- Semantic Search / FAQ Matching ---
sentence-transformers>=3.0.1
torch>=2.3.0
transformers>=4.44.0

# --- AWS SDK (Optional for audio uploads) ---
boto3>=1.34.0

# --- Utilities ---
pandas>=2.2.2
numpy>=1.26.4
difflib
fuzzywuzzy>=0.18.0
python-Levenshtein>=0.25.0

# --- Optional: Data Validation & Typing ---
pydantic>=2.11.7
typing-extensions>=4.12.2

# --- Miscellaneous (recommended for production) ---
tqdm>=4.66.4
colorama>=0.4.6
